Move first screen content
